reorganize windowing files together various windowing cleanup
