+++ src/lib.rs
@@ -67,7 +67,6 @@ pub mod project_files;
 pub mod project_select;
 pub mod reflect_default;
 pub mod remote;
-pub mod repo_link;
 pub mod restart;
 pub mod scene_io;
 pub mod scene_ops;
@@ -88,8 +87,7 @@ pub mod viewport;
 pub mod viewport_overlays;
 pub mod viewport_select;
 pub mod viewport_util;
-pub mod window_chrome;
-pub mod window_header;
+pub mod windowing;
 pub mod workspace_dropdown;
 
 use bevy::{
@@ -331,11 +329,7 @@ impl Plugin for EditorCorePlugin {
         .add_plugins(operator_tooltip::OperatorTooltipPlugin)
         .add_plugins(jackdaw_node_graph::NodeGraphPlugin)
         .add_plugins(jackdaw_animation::AnimationPlugin)
-        .add_plugins((
-            repo_link::RepoLinkPlugin,
-            window_chrome::WindowChromePlugin,
-            window_header::WindowHeaderPlugin,
-        ))
+        .add_plugins(windowing::WindowingPlugin)
         .add_plugins(jackdaw_panels::DockPlugin)
         .add_plugins(jackdaw_api_internal::ExtensionLoaderPlugin)
         .add_plugins(extension_watcher::ExtensionWatcherPlugin)
@@ -595,7 +589,7 @@ fn spawn_layout(
     mut commands: Commands,
     icon_font: Res<jackdaw_feathers::icons::IconFont>,
     editor_font: Res<jackdaw_feathers::icons::EditorFont>,
-    jackdaw_icon: Res<repo_link::JackdawIcon>,
+    jackdaw_icon: Res<windowing::JackdawIcon>,
 ) {
     commands.spawn((Camera2d, EditorEntity));
     commands.spawn(layout::editor_layout(
