+++ src/windowing/header.rs
@@ -6,13 +6,14 @@ use bevy::prelude::*;
 use bevy::window::{PrimaryWindow, Window};
 use jackdaw_feathers::tokens;
 
-use crate::{EditorEntity, repo_link, window_chrome};
+use crate::EditorEntity;
+
+use super::repo_link;
+
+use super::controls;
 
 #[derive(Component)]
 pub struct WindowHeaderRoot;
-
-#[derive(Component)]
-pub struct WindowHeaderDragRegion;
 
 pub struct WindowHeaderPlugin;
 
@@ -28,11 +29,11 @@ pub fn window_header(
     jackdaw_icon: Handle<Image>,
     content: impl Bundle,
 ) -> impl Bundle {
+    // order controls to the left on mac
     #[cfg(target_os = "macos")]
     let foreground_row = (
         EditorEntity,
         Pickable::IGNORE,
-        GlobalZIndex(1),
         Node {
             position_type: PositionType::Absolute,
             top: px(0.0),
@@ -54,7 +55,6 @@ pub fn window_header(
     let foreground_row = (
         EditorEntity,
         Pickable::IGNORE,
-        GlobalZIndex(1),
         Node {
             position_type: PositionType::Absolute,
             top: px(0.0),
@@ -79,30 +79,11 @@ pub fn window_header(
             position_type: PositionType::Relative,
             width: percent(100),
             height: px(tokens::WINDOW_HEADER_HEIGHT),
-            flex_shrink: 0.0,
             ..default()
         },
         BackgroundColor(tokens::WINDOW_BG),
-        BorderColor::all(tokens::BORDER_SUBTLE),
-        Pickable::IGNORE,
-        children![header_drag_layer(), foreground_row],
+        children![foreground_row],
     );
-}
-
-fn header_drag_layer() -> impl Bundle {
-    (
-        WindowHeaderDragRegion,
-        EditorEntity,
-        Node {
-            position_type: PositionType::Absolute,
-            top: px(0.0),
-            left: px(0.0),
-            width: percent(100),
-            height: percent(100),
-            ..default()
-        },
-        BackgroundColor(Color::srgba(0.0, 0.0, 0.0, 0.001)),
-    )
 }
 
 fn content_slot(content: impl Bundle) -> impl Bundle {
@@ -111,7 +92,6 @@ fn content_slot(content: impl Bundle) -> impl Bundle {
         Pickable::IGNORE,
         Node {
             flex_grow: 1.0,
-            flex_shrink: 1.0,
             min_width: px(0.0),
             height: percent(100),
             overflow: Overflow::clip(),
@@ -132,7 +112,6 @@ fn chrome_repo_slot(jackdaw_icon: Handle<Image>) -> impl Bundle {
         Pickable::IGNORE,
         Node {
             flex_shrink: 0.0,
-            flex_grow: 0.0,
             padding,
             height: percent(100),
             align_items: AlignItems::Center,
@@ -153,45 +132,28 @@ fn chrome_controls_slot(icon_font: Handle<Font>) -> impl Bundle {
         Pickable::IGNORE,
         Node {
             flex_shrink: 0.0,
-            flex_grow: 0.0,
             padding,
             height: percent(100),
             align_items: AlignItems::Center,
             ..default()
         },
-        children![window_chrome::window_controls(icon_font)],
+        children![controls::window_controls(icon_font)],
     );
 }
 
 fn on_drag_region_press(
     press: On<Pointer<Press>>,
-    regions: Query<Entity, With<WindowHeaderDragRegion>>,
-    parents: Query<&ChildOf>,
+    headers: Query<Entity, With<WindowHeaderRoot>>,
     mut windows: Query<&mut Window, With<PrimaryWindow>>,
 ) {
     if press.button != PointerButton::Primary {
         return;
     }
-    let Some(_region) = find_ancestor_with(press.event_target(), &regions, &parents) else {
+    if headers.get(press.original_event_target()).is_err() {
+        return;
+    }
+    let Ok(mut window) = windows.single_mut() else {
         return;
     };
-    for mut window in windows.iter_mut() {
-        window.start_drag_move();
-    }
-}
-
-fn find_ancestor_with(
-    mut entity: Entity,
-    query: &Query<Entity, With<WindowHeaderDragRegion>>,
-    parents: &Query<&ChildOf>,
-) -> Option<Entity> {
-    loop {
-        if query.contains(entity) {
-            return Some(entity);
-        }
-        let Ok(parent) = parents.get(entity) else {
-            return None;
-        };
-        entity = parent.parent();
-    }
+    window.start_drag_move();
 }
